Revert "DELIA-32699 : Changed rate parameter from float to int"
This reverts commit 7057cbbb853f7b6779f1380bc35299f5859168ae.

+++ main_aamp.h
@@ -634,7 +634,7 @@ public:
 	 *   @param[in]  overshoot - overshoot correction in milliseconds.
 	 *   @return void
 	 */
-	void SetRate(int rate, int overshootcorrection=0);
+	void SetRate(float rate, int overshootcorrection=0);
 
 	/**
 	 *   @brief Seek to a time.
@@ -657,7 +657,7 @@ public:
 	 *   @param[in]  secondsRelativeToTuneTime - Seek position for VOD,
 	 *           relative position from first tune command.
 	 */
-	void SetRateAndSeek(int rate, double secondsRelativeToTuneTime);
+	void SetRateAndSeek(float rate, double secondsRelativeToTuneTime);
 
 	/**
 	 *   @brief Register event handler.
@@ -925,7 +925,7 @@ public:
 	 *
 	 *   @ret current playback rate
 	 */
-	int GetPlaybackRate(void);
+	float GetPlaybackRate(void);
 
 	/**
 	 *   @brief To get the available video bitrates.
